fix: fix components impors and move routes to views
