+++ frontend/src/views/components/index.js
@@ -1,7 +1,6 @@
 import AppNavbar from './navbar'
-import AppRoutes from './routes'
 import CMSOMSCard from './cmsOMSCard'
 import ResponsivePlot from './responsivePlot'
 import Table from './table'
 
-export { AppNavbar, AppRoutes, CMSOMSCard, ResponsivePlot, Table }
+export { AppNavbar, CMSOMSCard, ResponsivePlot, Table }
